Add map to admin context for storing request data
This adds a `map[string]interface{}` to the admin context struct for
storing request specific data in a generic way. This way middleware is
able to store data for usage later in the chain without resorting to
storing it in the request (or some other inappropriate place).
This was copied over from Gin’s context struct.